Move resource and ibc parsing to a better location
