+++ src/ILCompiler.ReadyToRun/src/ILCompiler.ReadyToRun.csproj
@@ -155,9 +155,9 @@
     <Compile Include="Compiler\DependencyAnalysis\ReadyToRunSymbolNodeFactory.cs" />
     <Compile Include="Compiler\DependencyAnalysis\TypeAndMethod.cs" />
     <Compile Include="Compiler\IRootingServiceProvider.cs" />
-    <Compile Include="Compiler\ibcdatamodel.cs" />
-    <Compile Include="Compiler\ibcdatareader.cs" />
-    <Compile Include="Compiler\ibcextensions.cs" />
+    <Compile Include="IBC\ibcdatamodel.cs" />
+    <Compile Include="IBC\ibcdatareader.cs" />
+    <Compile Include="IBC\ibcextensions.cs" />
     <Compile Include="Compiler\ProfileData.cs" />
     <Compile Include="Compiler\ReadyToRunHashCode.cs" />
     <Compile Include="Compiler\ReadyToRunLibraryRootProvider.cs" />
@@ -167,14 +167,16 @@
     <Compile Include="Compiler\ReadyToRunMetadataFieldLayoutAlgorithm.cs" />
     <Compile Include="Compiler\ReadyToRunSingleAssemblyCompilationModuleGroup.cs" />
     <Compile Include="Compiler\ReadyToRunTableManager.cs" />
-    <Compile Include="Compiler\Resources\ResourceData.cs" />
-    <Compile Include="Compiler\Resources\ResourceData.FileAbstraction.cs" />
-    <Compile Include="Compiler\Resources\ResourceData.Reader.cs" />
-    <Compile Include="Compiler\Resources\ResourceData.ResourcesDataModel.cs" />
-    <Compile Include="Compiler\Resources\ResourceData.UpdateResourceDataModel.cs" />
-    <Compile Include="Compiler\Resources\ResourceData.Win32Enums.cs" />
-    <Compile Include="Compiler\Resources\ResourceData.Win32Structs.cs" />
-    <Compile Include="Compiler\Resources\ResourceData.Win32_NT_HEADER.cs" />
+    <Compile Include="IBC\IBCProfileData.cs" />
+    <Compile Include="IBC\IBCProfileParser.cs" />
+    <Compile Include="Win32Resources\ResourceData.cs" />
+    <Compile Include="Win32Resources\ResourceData.FileAbstraction.cs" />
+    <Compile Include="Win32Resources\ResourceData.Reader.cs" />
+    <Compile Include="Win32Resources\ResourceData.ResourcesDataModel.cs" />
+    <Compile Include="Win32Resources\ResourceData.UpdateResourceDataModel.cs" />
+    <Compile Include="Win32Resources\ResourceData.Win32Enums.cs" />
+    <Compile Include="Win32Resources\ResourceData.Win32Structs.cs" />
+    <Compile Include="Win32Resources\ResourceData.Win32_NT_HEADER.cs" />
     <Compile Include="Compiler\RuntimeDeterminedTypeHelper.cs" />
     <Compile Include="Compiler\SingleMethodCompilationModuleGroup.cs" />
     <Compile Include="Compiler\SystemObjectFieldLayoutAlgorithm.cs" />
